requirements.txt
numpy<2.0
torch
torchvision
Pillow
matplotlib
pathlib
tensorboard
torchsummary
scikit-learn